RTC: Re-add invalid data handling
This commit adds back some handling of invalid data that we had added to
LERC when it was embedded in MR3D. This fix isn't permanent, but should
work in the interim while we lack resources to do the more significant
update to using the LERC API properly. This commit inserts a parameter,
`bool& hasInvalidData`, which we use in runtime to help determine what
to do when there are multiple elevation data sources. Some may be
incomplete, and the handling of that is more delicate than we initially
thought, with the code added in this commit playing a critical role in
other parts of the MR3D codebase. To avoid confusing situations in which
the argument intended as `hasInvalidData` instead binds to `headerOnly`
after the update, this commit moves the `hasInvalidData` parameter over
to the left, so there's no ambiguity.

Signed-off-by: Nathan LaPre <nlapre@esri.com>